simplify scope based API
