+++ samples/Clients/Clients.sln
@@ -7,7 +7,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{187AC294-042
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{179A2A2C-3B3D-437C-B5F3-3B81472C8335}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "API", "API", "{AFE7085F-051E-4829-955F-3426FE643BDD}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "APIs", "APIs", "{AFE7085F-051E-4829-955F-3426FE643BDD}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleClientCredentialsFlow", "src\ConsoleClientCredentialsFlow\ConsoleClientCredentialsFlow.csproj", "{6729A51B-A7D9-43F3-8F48-D55478254AB7}"
 EndProject
@@ -28,8 +28,6 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleExtensionGrant", "src\ConsoleCustomGrant\ConsoleExtensionGrant.csproj", "{546267CD-BE21-44E1-A06E-C7558864EE4F}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JsOidc", "src\JsOidc\JsOidc.csproj", "{A29115B9-30E4-4967-AD5A-C32394B8A64C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ScopeBasedApi", "src\SampleApi\ScopeBasedApi.csproj", "{03D5E439-393F-4E03-BA8F-AD3462397B38}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleResourceOwnerFlowPublic", "src\ConsoleResourceOwnerFlowPublic\ConsoleResourceOwnerFlowPublic.csproj", "{8648122C-EE91-4B3B-98CD-355C3F1B9FDF}"
 EndProject
@@ -58,6 +56,8 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Console", "Console", "{D027D36B-262B-450A-B444-5B7893B5142E}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleParameterizedScopeClient", "src\ConsoleParameterizedScopeClient\ConsoleParameterizedScopeClient.csproj", "{AE8D301D-BDDD-4E04-A487-FD08BAF18B75}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleApi", "src\APIs\SimpleApi\SimpleApi.csproj", "{215D863E-9D87-42F4-AD17-209C330C6ECB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -105,10 +105,6 @@ Global
 		{A29115B9-30E4-4967-AD5A-C32394B8A64C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A29115B9-30E4-4967-AD5A-C32394B8A64C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A29115B9-30E4-4967-AD5A-C32394B8A64C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{03D5E439-393F-4E03-BA8F-AD3462397B38}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{03D5E439-393F-4E03-BA8F-AD3462397B38}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{03D5E439-393F-4E03-BA8F-AD3462397B38}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{03D5E439-393F-4E03-BA8F-AD3462397B38}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8648122C-EE91-4B3B-98CD-355C3F1B9FDF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8648122C-EE91-4B3B-98CD-355C3F1B9FDF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8648122C-EE91-4B3B-98CD-355C3F1B9FDF}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -157,6 +153,10 @@ Global
 		{AE8D301D-BDDD-4E04-A487-FD08BAF18B75}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AE8D301D-BDDD-4E04-A487-FD08BAF18B75}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AE8D301D-BDDD-4E04-A487-FD08BAF18B75}.Release|Any CPU.Build.0 = Release|Any CPU
+		{215D863E-9D87-42F4-AD17-209C330C6ECB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{215D863E-9D87-42F4-AD17-209C330C6ECB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{215D863E-9D87-42F4-AD17-209C330C6ECB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{215D863E-9D87-42F4-AD17-209C330C6ECB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -173,7 +173,6 @@ Global
 		{0E9145DB-CC70-44F9-B52B-1D7D4C2B2E5F} = {D027D36B-262B-450A-B444-5B7893B5142E}
 		{546267CD-BE21-44E1-A06E-C7558864EE4F} = {D027D36B-262B-450A-B444-5B7893B5142E}
 		{A29115B9-30E4-4967-AD5A-C32394B8A64C} = {158628D7-8B68-451E-AF22-B64F473C5943}
-		{03D5E439-393F-4E03-BA8F-AD3462397B38} = {AFE7085F-051E-4829-955F-3426FE643BDD}
 		{8648122C-EE91-4B3B-98CD-355C3F1B9FDF} = {D027D36B-262B-450A-B444-5B7893B5142E}
 		{1EC25581-781E-4FEA-A8AC-72BBD599FB87} = {D027D36B-262B-450A-B444-5B7893B5142E}
 		{BB38F47A-1BBF-47E1-9872-088FC81DC825} = {D027D36B-262B-450A-B444-5B7893B5142E}
@@ -188,6 +187,7 @@ Global
 		{158628D7-8B68-451E-AF22-B64F473C5943} = {187AC294-0420-4726-808B-ED49148C0632}
 		{D027D36B-262B-450A-B444-5B7893B5142E} = {187AC294-0420-4726-808B-ED49148C0632}
 		{AE8D301D-BDDD-4E04-A487-FD08BAF18B75} = {D027D36B-262B-450A-B444-5B7893B5142E}
+		{215D863E-9D87-42F4-AD17-209C330C6ECB} = {AFE7085F-051E-4829-955F-3426FE643BDD}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {BAD78470-3D66-466E-9C17-2A67F0905B18}
