+++ samples/Clients/Directory.Build.targets
@@ -5,11 +5,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Update="IdentityModel" Version="4.1.0" />
-    <PackageReference Update="IdentityServer4.AccessTokenValidation" Version="3.0.1" />
-
+    <PackageReference Update="IdentityModel" Version="4.3.0" />
+    
     <PackageReference Update="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="$(AspNetVersion)" />
     <PackageReference Update="Microsoft.AspNetCore.Authentication.OpenIdConnect" Version="$(AspNetVersion)" />
+    <PackageReference Update="Microsoft.AspNetCore.Authentication.JwtBearer" Version="$(AspNetVersion)" />     
     <PackageReference Update="Microsoft.AspNetCore.Authentication.Certificate" Version="$(AspNetVersion)"/>
     
     <PackageReference Update="System.IdentityModel.Tokens.Jwt" Version="5.6.0" />
