+++ samples/Clients/src/APIs/SimpleApi/Program.cs
