+++ samples/Clients/src/APIs/SimpleApi/Properties/launchSettings.json
