+++ samples/Clients/src/APIs/SimpleApi/SimpleApi.csproj
@@ -5,14 +5,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Constants\Constants.csproj" />
+    <ProjectReference Include="..\..\Constants\Constants.csproj" />
   </ItemGroup>
   
   <ItemGroup>
-    <PackageReference Include="IdentityServer4.AccessTokenValidation" />
-    
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.Certificate" />
     <PackageReference Include="System.IdentityModel.Tokens.Jwt" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" />  
     
     <PackageReference Include="Serilog.AspNetCore" Version="3.2.0" />
   </ItemGroup>
