+++ samples/Clients/src/APIs/SimpleApi/Startup.cs
@@ -1,0 +1,56 @@
+﻿using Clients;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.DependencyInjection;
+using System;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Cryptography.X509Certificates;
+using System.Text;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.HttpOverrides;
+
+namespace SampleApi
+{
+    public class Startup
+    {
+        public void ConfigureServices(IServiceCollection services)
+        {
+            services.AddControllers();
+
+            services.AddCors();
+            services.AddDistributedMemoryCache();
+
+            // this API will accept any access token from the authority
+            services.AddAuthentication("token")
+                .AddJwtBearer("token", options =>
+                {
+                    options.Authority = Constants.Authority;
+                    options.TokenValidationParameters.ValidateAudience = false;
+                    
+                    options.TokenValidationParameters.ValidTypes = new[] { "at+jwt" };
+                });
+        }
+
+        public void Configure(IApplicationBuilder app)
+        {
+            app.UseCors(policy =>
+            {
+                policy.WithOrigins(
+                    "https://localhost:44300");
+
+                policy.AllowAnyHeader();
+                policy.AllowAnyMethod();
+                policy.WithExposedHeaders("WWW-Authenticate");
+            });
+
+            app.UseRouting();
+            app.UseAuthentication();
+            app.UseAuthorization();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers().RequireAuthorization();
+            });
+        }
+    }
+}
