--- samples/Clients/src/SampleApi/web.config
@@ -1,20 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <system.webServer>
-    <handlers>
-      <add name="aspNetCore" path="*" verb="*" modules="AspNetCoreModuleV2" resourceType="Unspecified" />
-    </handlers>
-    <aspNetCore processPath="%LAUNCHER_PATH%" arguments="%LAUNCHER_ARGS%" stdoutLogEnabled="false" hostingModel="InProcess">
-      <environmentVariables>
-        <environmentVariable name="ASPNETCORE_ENVIRONMENT" value="Development" />
-      </environmentVariables>
-    </aspNetCore>
-  </system.webServer>
-  <!--<location path="identity">
-    <system.webServer>
-      <security>
-        <access sslFlags="Ssl, SslNegotiateCert, SslRequireCert" />
-      </security>
-    </system.webServer>
-  </location>-->
-</configuration>
